+++ src/ApolloClient.ts
@@ -44,6 +44,7 @@ import {
 
 import {
   ApolloQueryResult,
+  ApolloExecutionResult,
   IdGetter,
 } from './core/types';
 
@@ -357,7 +358,7 @@ export default class ApolloClient implements DataProxy {
    *
    * It takes options as an object with the following keys and values:
    */
-  public mutate<T>(options: MutationOptions): Promise<ExecutionResult> {
+  public mutate<T>(options: MutationOptions<T>): Promise<ApolloExecutionResult<T>> {
     this.initStore();
 
     return this.queryManager.mutate<T>(options);