+++ src/core/types.ts
@@ -1,4 +1,4 @@
-import { DocumentNode } from 'graphql';
+import { DocumentNode, GraphQLError } from 'graphql';
 import { QueryStoreValue } from '../queries/store';
 import { NetworkStatus } from '../queries/networkStatus';
 
@@ -15,8 +15,16 @@ export type ApolloQueryResult<T> = {
   networkStatus: NetworkStatus;
   stale: boolean;
 
-  // This type is different from the GraphQLResult type because it doesn't include errors.
+  // This type is different from the ExecutionResult type because it doesn't include errors.
   // Those are thrown via the standard promise/observer catch mechanism.
+};
+
+export type ApolloExecutionResult<T = { [key: string]: any }> = {
+  data?: T;
+
+  // This type is different from the ExecutionResult type because it doesn't include errors.
+  // Those are thrown via the standard promise/observer catch mechanism.
+  // It also has a generic type
 };
 
 export enum FetchType {